Add files via upload
Recognizing textual entailment (RTE) is a task that predict whether a text fragment can be inferred from another text fragment. On the training data we have extracted textual features and have trained the classification model (SVM & Random forest).  Secondly, the developed model was used to predict entailment labels on the test data. To ensure scalability the project was created in PySpark. One function for feature extraction was created, and the execution of this function was parallelized. Both training and testing data consist of 800 text-hypothesis pairs.

For each text-hypothesis pair the following features were extracted:
Word overlap (the number of similar words in the text and in the hypothesis)
Bigram overlap (the number of similar bigrams in the text and in the hypothesis)
Trigram overlap (the number of similar trigrams in the text and in the hypothesis)
Nouns overlap
Proper nouns overlap
Verb overlap
Adverb overlap
Adjective overlap
Number, date, time (numeric) featu